+++ include/proxsuite/linalg/veg/internal/fix_index.hpp
@@ -326,7 +326,7 @@ namespace adl {
 inline namespace literals {
 template<char... Chars>
 VEG_INLINE constexpr auto
-operator"" _c() VEG_NOEXCEPT
+operator""_c() VEG_NOEXCEPT
 {
   return Fix<_detail::parse_int(
     _detail::char_seq<Chars...>::value, sizeof...(Chars), _detail::Error{})>{};